Revert "Cloned the project to repo."
This reverts commit d578cb33076fec5b96e49d58342bd84a7c3694f2.
