+++ src/main/java/ru/skypro/homework/controller/BasicAuthCorsFilter.java
@@ -1,4 +1,4 @@
-package ru.skypro.homework.filter;
+package ru.skypro.homework.controller;
 
 
 import org.springframework.stereotype.Component;
