+++ src/main/java/ru/skypro/homework/dto/LoginReq.java
@@ -3,8 +3,8 @@ package ru.skypro.homework.dto;
 import lombok.Data;
 
 @Data
-public class Login {
-
-    private String username;
+public class LoginReq {
     private String password;
+    private String username;
+
 }
