+++ src/main/java/ru/skypro/homework/dto/RegisterReq.java
@@ -3,8 +3,7 @@ package ru.skypro.homework.dto;
 import lombok.Data;
 
 @Data
-public class Register {
-
+public class RegisterReq {
     private String username;
     private String password;
     private String firstName;
